requirements.txt
langchain
chromadb
flask
flask_cors
huggingface_hub
qdrant-client
python-dotenv
sentence_transformers